+++ CHANGES.md
@@ -8,6 +8,7 @@ To be released.
 
  -  Since this version, it requires GHC 8.8.* at least, and supports GHC 9.0.*
     at most.
+ -  Added `Text.Seonbi.Html.Lang` module.
  -  Removed several functions from `Text.Seonbi.Trie` module:
      -  `toListBy`
      -  `lookupBy`